+++ src/jtl/Connector/Shopware/Controller/CustomerOrder.php
@@ -302,8 +302,11 @@ class CustomerOrder extends DataController
                     }
 
                     // Update order status
-                    $value = Application()->getConfig()->get('customer_order_processing_after_pull');
-                    if ($swOrderObj->getOrderStatus()->getId() === Status::ORDER_STATE_OPEN && (is_null($value) || $value === true)) {
+
+                    /** @deprecated Will be removed in a future connector release $valueOld */
+                    $valueOld = Application()->getConfig()->get('customer_order_processing_after_pull', true);
+                    $value = Application()->getConfig()->get('customer_order.pull.status_processing', $valueOld);
+                    if ($swOrderObj->getOrderStatus()->getId() === Status::ORDER_STATE_OPEN && $value) {
                         $status = Shopware()->Models()->getRepository('Shopware\Models\Order\Status')->findOneById(Status::ORDER_STATE_IN_PROCESS);
 
                         $swOrderObj->setOrderStatus($status);